Update conf.py
The API in cosipy docs is missing dataIO, Polarization, and Utilities, meaning it didn't successfully compile in the Spinx build. Checking the logs in the cosipy Actions tab, I see that the failure seems to be from the missing package: iminuit. So I added this to autodoc_mock_imports, and hopefully that fixes things.